requirements.txt
ttkbootstrap
Pillow
PyAudio
SpeechRecognition
googletrans
gradio-client
deep-translator